Optimise building Trans value, storage of static eval

+++ EubosChess/src/main/java/eubos/search/PlySearcher.java
@@ -708,7 +708,7 @@ public class PlySearcher {
 		if (Score.isMate(plyScore)) {
 			scoreFromDownTree = (short) ((plyScore < 0) ? plyScore - currPly : plyScore + currPly);
 		}
-		trans = tt.setTransposition(pos.getHash(), trans, depth, scoreFromDownTree, plyBound, currMove, pos.getMoveNumber(), 
+		trans = tt.setTransposition(pos.getHash(), trans, depth, scoreFromDownTree, plyBound, (short)currMove, pos.getMoveNumber(), 
 				s.isStaticValid ? s.staticEval : Short.MAX_VALUE);
 		return trans;
 	}
