+++ EubosChess/src/main/java/eubos/search/transposition/DummyTranspositionTable.java
@@ -16,7 +16,7 @@ public class DummyTranspositionTable implements ITranspositionAccessor {
 	public synchronized void putTransposition(long hashCode, long trans) {
 	}
 	
-	public long setTransposition(long hash, long trans, byte new_Depth, short new_score, byte new_bound, int new_bestMove, int move_number, short static_eval) {
+	public long setTransposition(long hash, long trans, byte new_Depth, short new_score, byte new_bound, short new_bestMove, int move_number, short static_eval) {
 		return 0L;
 	}
